refactor: directory refactor
