+++ benches/benchmark_e448_512.rs
@@ -1,6 +1,6 @@
 use capycrypt::ecc::ecc_encryptable::KeyEncryptable;
+use capycrypt::ecc::ecc_keypair::KeyPair;
 use capycrypt::ecc::ecc_signable::Signable;
-use capycrypt::ecc::keypair::KeyPair;
 use capycrypt::sha3::aux_functions::byte_utils::get_random_bytes;
 use capycrypt::sha3::sponge_encryptable::SpongeEncryptable;
 use capycrypt::SecParam::D512;
