+++ benches/benchmark_sha3.rs
@@ -9,7 +9,7 @@ const BIT_SECURITY: SecParam = D256;
 
 /// hash 5mb of random data with 128 bits of security
 fn sha3_digest(mut msg: Message) {
-    let _ = msg.compute_hash_sha3(&BIT_SECURITY);
+    let _ = msg.compute_sha3_hash(&BIT_SECURITY);
 }
 
 fn bench_sha3_digest(c: &mut Criterion) {
