--- src/ecc/mod.rs
@@ -1,3 +1,0 @@
-pub mod ecc_encryptable;
-pub mod ecc_signable;
-pub mod keypair;
