+++ src/kem/kem_encryptable.rs
@@ -7,38 +7,19 @@ use crate::{
 };
 use capy_kem::{
     constants::parameter_sets::KEM_768,
-    fips203::{decrypt::k_pke_decrypt, encrypt::k_pke_encrypt, keygen::k_pke_keygen},
+    fips203::{decrypt::k_pke_decrypt, encrypt::k_pke_encrypt},
 };
 use rand::{thread_rng, RngCore};
-use serde::{Deserialize, Serialize};
+
+use super::kem_keypair::{KEMPrivateKey, KEMPublicKey};
 
 pub trait KEMEncryptable {
-    fn kem_encrypt(&mut self, key: &KEMKey, d: &SecParam) -> Result<(), OperationError>;
-    fn kem_decrypt(&mut self, key: &KEMKey) -> Result<(), OperationError>;
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct KEMKey {
-    pub ek: Vec<u8>,
-    pub dk: Vec<u8>,
-    pub rand_bytes: [u8; 32],
-}
-
-impl KEMKey {
-    pub fn kem_keygen() -> KEMKey {
-        let mut rng = thread_rng();
-        let mut rand_bytes = [0u8; 32];
-
-        // generate randomness for the KEM
-        rng.fill_bytes(&mut rand_bytes);
-        let (ek, dk) = k_pke_keygen::<KEM_768>(&rand_bytes);
-
-        KEMKey { ek, dk, rand_bytes }
-    }
+    fn kem_encrypt(&mut self, key: &KEMPublicKey, d: &SecParam) -> Result<(), OperationError>;
+    fn kem_decrypt(&mut self, key: &KEMPrivateKey) -> Result<(), OperationError>;
 }
 
 impl KEMEncryptable for Message {
-    fn kem_encrypt(&mut self, key: &KEMKey, d: &SecParam) -> Result<(), OperationError> {
+    fn kem_encrypt(&mut self, key: &KEMPublicKey, d: &SecParam) -> Result<(), OperationError> {
         self.d = Some(*d);
 
         let mut rng = thread_rng();
@@ -66,7 +47,7 @@ impl KEMEncryptable for Message {
         Ok(())
     }
 
-    fn kem_decrypt(&mut self, key: &KEMKey) -> Result<(), OperationError> {
+    fn kem_decrypt(&mut self, key: &KEMPrivateKey) -> Result<(), OperationError> {
         let ciphertext = self
             .kem_ciphertext
             .as_ref()
