--- src/kem/mod.rs
@@ -1,1 +1,0 @@
-pub mod kem_encryptable;
